+++ js/const.js
@@ -1,0 +1,44 @@
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо.',
+  'Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.',
+  'В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.',
+  'Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Александр',
+  'Екатерина',
+  'Иван',
+  'Анна',
+  'Михаил',
+  'Ольга',
+  'Сергей',
+  'Наталья',
+  'Дмитрий',
+  'Мария',
+];
+
+const DESCRIPTION = [
+  'Природная красота во всей своей палитре.',
+  'Момент вечной гармонии.',
+  'Искусство в природе.',
+  'Захватывающая игра света и тени.',
+  'Симфония природы.',
+  'Величие момента.',
+  'Природное великолепие.',
+  'Эмоциональное отражение.',
+  'Сказочная атмосфера.',
+  'Легендарная красота.',
+  'Вдохновение искусством природы.',
+  'Эффектное преображение.',
+  'Стремление к идеалу.',
+  'Загадочная глубина.',
+  'Мир, окутанный тайной.',
+];
+
+export {MESSAGE, NAMES, DESCRIPTION};
